+++ cargo-insta/src/cli.rs
@@ -18,13 +18,13 @@ use structopt::StructOpt;
 use uuid::Uuid;
 
 use crate::cargo::{
-    find_packages, find_snapshots, get_cargo, get_package_metadata, Operation, Package,
+    find_packages, find_snapshots, get_cargo, get_package_metadata, FindFlags, Operation, Package,
     SnapshotContainer,
 };
 use crate::utils::{err_msg, QuietExit};
 
-const IGNORE_MESSAGE: &str =
-    "some paths are ignored, use --no-ignore if you have snapshots in ignored paths.";
+const INCLUDE_IGNORED_HINT: &str =
+    "some paths are ignored, use --include-ignored if you have snapshots in ignored paths.";
 
 /// A helper utility to work with insta snapshots.
 #[derive(StructOpt, Debug)]
@@ -87,8 +87,11 @@ pub struct TargetArgs {
     #[structopt(long)]
     pub all: bool,
     /// Also walk into ignored paths.
+    #[structopt(long, alias = "no-ignore")]
+    pub include_ignored: bool,
+    /// Also include hidden paths.
     #[structopt(long)]
-    pub no_ignore: bool,
+    pub include_hidden: bool,
 }
 
 #[derive(StructOpt, Debug)]
@@ -310,7 +313,7 @@ struct LocationInfo<'a> {
     workspace_root: PathBuf,
     packages: Option<Vec<Package>>,
     exts: Vec<&'a str>,
-    no_ignore: bool,
+    find_flags: FindFlags,
 }
 
 fn handle_target_args(target_args: &TargetArgs) -> Result<LocationInfo<'_>, Box<dyn Error>> {
@@ -344,13 +347,17 @@ fn handle_target_args(target_args: &TargetArgs) -> Result<LocationInfo<'_>, Box<
         (None, None) => (None, None),
     };
 
+    let find_flags = FindFlags {
+        include_ignored: target_args.include_ignored,
+        include_hidden: target_args.include_hidden,
+    };
     if let Some(workspace_root) = workspace_root {
         Ok(LocationInfo {
             tool_config: ToolConfig::from_workspace(&workspace_root)?,
             workspace_root: workspace_root.to_owned(),
             packages: None,
             exts,
-            no_ignore: target_args.no_ignore,
+            find_flags,
         })
     } else {
         let metadata = get_package_metadata(manifest_path.as_ref().map(|x| x.as_path()))?;
@@ -360,7 +367,7 @@ fn handle_target_args(target_args: &TargetArgs) -> Result<LocationInfo<'_>, Box<
             workspace_root: metadata.workspace_root().to_path_buf(),
             packages: Some(packages),
             exts,
-            no_ignore: target_args.no_ignore,
+            find_flags,
         })
     }
 }
@@ -372,7 +379,8 @@ fn load_snapshot_containers<'a>(
     match loc.packages {
         Some(ref packages) => {
             for package in packages.iter() {
-                for snapshot_container in package.iter_snapshot_containers(&loc.exts, loc.no_ignore)
+                for snapshot_container in
+                    package.iter_snapshot_containers(&loc.exts, loc.find_flags)
                 {
                     snapshot_containers.push((snapshot_container?, Some(package)));
                 }
@@ -380,7 +388,7 @@ fn load_snapshot_containers<'a>(
         }
         None => {
             for snapshot_container in
-                find_snapshots(loc.workspace_root.clone(), &loc.exts, loc.no_ignore)
+                find_snapshots(loc.workspace_root.clone(), &loc.exts, loc.find_flags)
             {
                 snapshot_containers.push((snapshot_container?, None));
             }
@@ -404,8 +412,12 @@ fn process_snapshots(
     if snapshot_count == 0 {
         if !quiet {
             println!("{}: no snapshots to review", style("done").bold());
-            if !loc.no_ignore {
-                println!("{}: {}", style("warning").yellow().bold(), IGNORE_MESSAGE);
+            if !loc.find_flags.include_ignored {
+                println!(
+                    "{}: {}",
+                    style("warning").yellow().bold(),
+                    INCLUDE_IGNORED_HINT
+                );
             }
         }
         return Ok(());
@@ -701,15 +713,16 @@ fn test_run(mut cmd: TestCommand, color: &str) -> Result<(), Box<dyn Error>> {
                 style(snapshot_count).yellow(),
                 if snapshot_count != 1 { "s" } else { "" }
             );
-            if !loc.no_ignore {
-                println!("      {}", IGNORE_MESSAGE);
-            }
             eprintln!("use `cargo insta review` to review snapshots");
             return Err(QuietExit(1).into());
         } else {
             println!("{}: no snapshots to review", style("info").bold());
-            if !loc.no_ignore {
-                println!("{}: {}", style("warning").yellow().bold(), IGNORE_MESSAGE);
+            if !loc.find_flags.include_ignored {
+                println!(
+                    "{}: {}",
+                    style("warning").yellow().bold(),
+                    INCLUDE_IGNORED_HINT
+                );
             }
         }
     }